pyproject.toml
[project]
name = "ugts-dti"
version = "0.1.0"
description = "UGTS-DTI: Uncertainty-Gated Teacher–Student Learning for Drug–Target Interaction Prediction"
readme = "README.md"
requires-python = ">=3.9,<3.11"
dependencies = [
    "torch>=1.10.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "loguru>=0.5.3",
    "pyyaml>=6.0",
    "deeppurpose>=0.1.5",
    "matplotlib>=3.4.0",
    "tqdm>=4.62.0",
    "rdkit-pypi>=2021.3.4",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[dependency-groups]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pre-commit>=3.8",
    "types-PyYAML>=6.0",
]

[tool.uv]
package = true

# ----------------------
# Ruff
# ----------------------
[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py310"
exclude = [
    ".venv",
    ".ruff_cache",
    ".mypy_cache",
    "output",
    "data",
]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort
  "UP", # pyupgrade
  "B",  # flake8-bugbear
  "SIM",# flake8-simplify
]
ignore = [
  "E501", # line length handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ----------------------
# Mypy
# ----------------------
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "DeepPurpose.*",
    "rdkit.*",
    "pandas.*",
    "numpy.*",
    "sklearn.*",
    "scipy.*",
    "matplotlib.*",
    "tqdm.*",
    "loguru.*",
    "yaml.*",
]
ignore_missing_imports = true
